+++ DearPyGui/src/core/AppItems/plots/mvAreaSeries.cpp
@@ -157,7 +157,7 @@ namespace Marvel {
 		xptr = &(*m_value.get())[0];
 		yptr = &(*m_value.get())[1];
 
-		ImPlot::PlotLine(m_label.c_str(), xptr->data(), yptr->data(), xptr->size());
+		ImPlot::PlotLine(m_label.c_str(), xptr->data(), yptr->data(), (int)xptr->size());
 
 		ImPlot::PushPlotClipRect();
 		auto item = ImPlot::RegisterOrGetItem(m_label.c_str());
